+++ generic-ec/CHANGELOG.md
@@ -1,3 +1,9 @@
+## v0.2.3
+* Add `generic_ec::serde::PreferCompact` that serializes points/scalars in compact form,
+  but deserialization recognizes both compact and non-compact formats [#28]
+
+[#28]: https://github.com/dfns/generic-ec/pull/28
+
 ## v0.2.2
 * Implement `serde_with::SerializeAs<&T>` for `generic_ec::serde::Compact` when `T` is
   serializable via `Compact` [#27]